Moved Driver Managers to src/main/java.
