+++ src/test/java/io/saucelabs/portal/qa/commons/BaseTest.java
@@ -1,7 +1,8 @@
 package io.saucelabs.portal.qa.commons;
 
-import java.lang.reflect.Method;
-
+import io.saucelabs.portal.qa.drivermanager.DriverManager;
+import io.saucelabs.portal.qa.drivermanager.WebDriverManager;
+import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.WebDriver;
 import org.testng.ITestContext;
 import org.testng.ITestResult;
@@ -9,7 +10,7 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeSuite;
 
-import lombok.extern.slf4j.Slf4j;
+import java.lang.reflect.Method;
 
 @Slf4j
 public abstract class BaseTest {
